requirements.txt
APScheduler==3.6.3
asgiref==3.6.0
cachetools==4.2.2
certifi==2022.12.7
Django==4.1.7
python-decouple==3.7
python-telegram-bot==13.13
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tzdata==2022.7
tzlocal==4.2
